feat: 주소록 확장 및 예외 처리 전략 추가
--- 본문 ---
- addresses 테이블에 alias 컬럼을 추가하고, user_id, postal_code, address_line에 대한 고유 인덱스를 생성했습니다.
- last_used_at 컬럼을 추가하여 마지막 사용 시간을 기록하도록 변경했습니다.
- BindException에 대한 예외 처리 전략을 추가하여 유효성 검사 오류 응답을 개선했습니다.
- Querydsl을 위한 JPAQueryFactory 빈 설정을 추가했습니다.
- PageResponse의 totalElements 타입을 long으로 변경했습니다.

+++ src/main/java/io/github/metdaisy/amaazon/common/dto/PageResponse.java
@@ -2,6 +2,6 @@ package io.github.metdaisy.amaazon.common.dto;
 
 import java.util.List;
 
-public record PageResponse<T>(List<T> data, int page, int size, int totalElements, int totalPages) {
+public record PageResponse<T>(List<T> data, int page, int size, long totalElements, int totalPages) {
 
 }
